Feature/streaming (#2)
- Streaming implementation, using gRPC
- Remove OwnedHashTreeNode
- Add Skipped variant to StoredChunkRef
- Remove `transfer` endpoint from REST API
- Switch to u64 for sizes
- Implement serialization for StoredChunkRef, node's children are
  serialized as a struct with hash and size only, on deserialization
  they are rebuilt as a Skipped node and need to be filled from storage
  before they can be used
- Make diff_tree() streaming, the new one should work nicely with gRPC
  while copying less data than before. It's far from zero-copy but
  certainly an improvement
- Fix HashMapStorage size method
-  streaming insertion to storage
- Remove interior mutability for ChunkStorage and both implementations
- Ignore some dirs and files used for manual testing
- Replace some locks server-side with async versions (I'm not sure it's
the right thing, but for now it's one thing less to keep and eye on).

Chunking and hash-tree node handling is very expensive, reducing
CHUNK_SIZE causes a *BIG* overhead. Even when everything is ironed out
and working fine, in real world use the size shouldn't be too small in
order to get the performance out of this transfer-chunks-thing.

+++ core/src/hash.rs
@@ -30,7 +30,7 @@ pub fn hash(data: &[u8]) -> hash::Hash {
         //println!("[HASH-NE] HASH: {}", x);
     }
 
-    let chunks: Vec<&[u8]> = data.chunks(CHUNK_SIZE).collect();
+    let chunks: Vec<&[u8]> = data.chunks(CHUNK_SIZE as usize).collect();
     partial_tree(&chunks)
 }
 
@@ -42,12 +42,13 @@ pub mod hash {
     use std::fmt;
 
     use blake3::OUT_LEN;
+    use serde::{Deserialize, Serialize};
 
     /// Reimplementation of blake3::Hash without constant-time comparisons
     ///
     /// Constant-time equality check is not needed for this use-case and only hurts performance
     /// Code is almost identical to blake3::Hash
-    #[derive(Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
+    #[derive(Clone, Copy, Serialize, Deserialize, Hash, PartialEq, Eq, PartialOrd, Ord)]
     pub struct Hash([u8; OUT_LEN]);
 
     impl Hash {
